Move retriever to be overridable
