+++ retriever_file_test.go
@@ -1,9 +1,9 @@
-package retriever_test
+package ffclient_test
 
 import (
 	"context"
 	"github.com/stretchr/testify/assert"
-	"github.com/thomaspoignant/go-feature-flag/internal/retriever"
+	ffclient "github.com/thomaspoignant/go-feature-flag"
 	"testing"
 )
 
@@ -35,7 +35,7 @@ func Test_localRetriever_Retrieve(t *testing.T) {
 		{
 			name: "File exists",
 			fields: fields{
-				path: "../../testdata/flag-config.yaml",
+				path: "./testdata/flag-config.yaml",
 			},
 			want:    []byte(expectedFile),
 			wantErr: false,
@@ -43,7 +43,7 @@ func Test_localRetriever_Retrieve(t *testing.T) {
 		{
 			name: "File does not exists",
 			fields: fields{
-				path: "../../testdata/test-not-exist.yaml",
+				path: "./testdata/test-not-exist.yaml",
 			},
 			want:    nil,
 			wantErr: true,
@@ -51,7 +51,7 @@ func Test_localRetriever_Retrieve(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			l := retriever.NewLocalRetriever(tt.fields.path)
+			l := ffclient.FileRetriever{Path: tt.fields.path}
 			got, err := l.Retrieve(context.Background())
 			if tt.wantErr {
 				assert.Error(t, err, "Retrieve() error = %v, wantErr %v", err, tt.wantErr)
