+++ testutils/mock_s3manager.go
@@ -17,7 +17,7 @@ type S3ManagerMock struct {
 func (s *S3ManagerMock) Download(at io.WriterAt, input *s3.GetObjectInput,
 	f ...func(*s3manager.Downloader)) (int64, error) {
 	if *input.Key == "valid" {
-		res, _ := ioutil.ReadFile("../../testdata/flag-config.yaml")
+		res, _ := ioutil.ReadFile("./testdata/flag-config.yaml")
 		_, _ = at.WriteAt(res, 0)
 		return 1, nil
 	} else if *input.Key == "no-file" {
